% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encryptionmatrices.R
\name{B.from_blocks}
\alias{B.from_blocks}
\title{Build encryption matrix from blocks of 2x2 matrices}
\usage{
B.from_blocks(a, b, c, d)
}
\arguments{
\item{a}{- 2x2 matrix}

\item{b}{- 2x2 matrix}

\item{c}{- 2x2 matrix}

\item{d}{- 2x2 matrix}
}
\value{
4x4 matrix
}
\description{
Build encryption matrix from blocks of 2x2 matrices
}
