% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaryconverters.R
\name{bcd_to_decimal}
\alias{bcd_to_decimal}
\title{Binary Code Decimal to Decimal code conversion}
\usage{
bcd_to_decimal(bcd)
}
\arguments{
\item{bcd}{- A vector of numbers in BCD}
}
\value{
Given number converted in decimal
}
\description{
Binary Code Decimal to Decimal code conversion
}
\examples{
bcd_number <- c(0, 1, 1, 1, 1, 0, 0, 0)
bcd_to_decimal(bcd_number)  # returns 78

}
\seealso{
\link{decimal_to_bcd}, \link{bcd_to_gray}
}
