% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaryconverters.R
\name{binary_to_decimal}
\alias{binary_to_decimal}
\title{Binary to decimal convertion}
\usage{
binary_to_decimal(binary)
}
\arguments{
\item{binary}{- Number in binary}
}
\value{
binary number converted to decimal
}
\description{
Binary to decimal convertion
}
\examples{
number <- c(1, 0, 1, 1)
binary_to_decimal(number)  # returns 7

}
\seealso{
\link{decimal_to_binary}, \link{binary_to_gray}
}
