% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaryconverters.R
\name{binary_to_gray}
\alias{binary_to_gray}
\title{Binary number to Gray number conversion}
\usage{
binary_to_gray(number)
}
\arguments{
\item{number}{- number in binary format}
}
\value{
Given binary nunmber in gray format
}
\description{
Binary number to Gray number conversion
}
\examples{
binary_three <- c(1,1)
binary_to_gray(binary_three)   # returns c(1,0)

}
\seealso{
\link{gray_to_binary}, \link{binary_to_decimal}
}
