% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codetable.R
\name{codetable1.characterise_num.from_matrix}
\alias{codetable1.characterise_num.from_matrix}
\title{Characterise numerals from a matrix}
\usage{
codetable1.characterise_num.from_matrix(numerals)
}
\arguments{
\item{numerals}{- Matrix of numerals}
}
\value{
raveled messaged containing characters
}
\description{
Characterise numerals from a matrix
}
\examples{
numerals <- array(
    c(8, 52, 31, 30, 30, 37, 52, 43, 30, 26, 29, 50, 54, 54, 54, 54),
    dim = c(4,4)
)
characters <- codetable1.characterise_num(numerals)
sentence(characters)
}
\seealso{
\link{codetable1.numerise_char}, \link{codetable1.characterise_num}
}
