% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codetable.R
\name{codetable1.characterise_num}
\alias{codetable1.characterise_num}
\title{Convert numerals to characters following code table}
\usage{
codetable1.characterise_num(num)
}
\arguments{
\item{num}{- Numerals to be converted}
}
\value{
Vector containing characters
}
\description{
Convert numerals to characters following code table
}
\examples{
numerals <- c(8, 45, 52, 34, 44, 52,  4, 37, 30, 28, 45, 43, 34, 28, 52, 53)
converted <- codetable1.characterise_num(numerals)
converted    # Converted from numerals to character
paste(converted, collapse="")  # Merge all

}
\seealso{
\link{codetable1.numerise_char}, \link{codetable1.characterise_num.from_matrix}
}
