% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codetable.R
\name{codetable1.numerise_char.as_matrix}
\alias{codetable1.numerise_char.as_matrix}
\title{Numerise character and convert it into a 4x4 matrix filled by rows}
\usage{
codetable1.numerise_char.as_matrix(char)
}
\arguments{
\item{char}{- Text to be converted}
}
\value{
4x4 Matrix containing the numerised char
}
\description{
Numerise character and convert it into a 4x4 matrix filled by rows
}
\examples{
codetable1.numerise_char.as_matrix("Quantum Mechanic")

}
\seealso{
\link{codetable1.numerise_char}
}
