% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codetable.R
\name{codetable1.numerise_char}
\alias{codetable1.numerise_char}
\title{Convert a character text to numerals following the code table}
\usage{
codetable1.numerise_char(char)
}
\arguments{
\item{char}{- Text to be converted}
}
\value{
Numbers representing the given text character in the code table
}
\description{
Convert a character text to numerals following the code table
}
\examples{
codetable1.numerise_char("Hello You")

}
\seealso{
\link{codetable1.numerise_char.as_matrix}
}
