% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaryconverters.R
\name{decimal_to_bcd}
\alias{decimal_to_bcd}
\title{Decimal number conversion to Binary Code Decimal}
\usage{
decimal_to_bcd(number)
}
\arguments{
\item{number}{- Can be a decimal number or vector of unit numbers representing digits of the number}
}
\value{
Given number in BCD
}
\description{
Decimal number conversion to Binary Code Decimal
}
\examples{
decimal_to_bcd(23)       # returns 0010 0101
decimal_to_bcd(c(2, 3))  # returns 0010 0101

}
\seealso{
\link{bcd_to_decimal}, \link{bcd_to_gray}
}
