% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaryconverters.R
\name{decimal_to_binary}
\alias{decimal_to_binary}
\title{Decimal to Binary conversion}
\usage{
decimal_to_binary(number)
}
\arguments{
\item{number}{- number in decimal}
}
\value{
number in binary format
}
\description{
Decimal to Binary conversion
}
\examples{
a <- 89
decimal_to_binary(a)   # Converts 89 to binary
decimal_to_binary(10)  # Output 1 0 1 0

}
\seealso{
\link{binary_to_decimal}, \link{decimal_to_bcd}
}
