% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaryconverters.R
\name{decimal_to_gray}
\alias{decimal_to_gray}
\title{Decimal number to gray code conversion}
\usage{
decimal_to_gray(decimal)
}
\arguments{
\item{decimal}{- A number in decimal format}
}
\value{
Given number in gray code
}
\description{
Decimal number to gray code conversion
}
\examples{
decimal_to_gray(5)   # return c(1,1,1)
}
\seealso{
{\link{gray_to_binary}}, {\link{gray_to_decimal}}
}
