% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definestate.R
\name{definestate}
\alias{definestate}
\title{define_state

load_all("./")}
\usage{
definestate(a, state)
}
\description{
define_state

load_all("./")
}
\examples{
# to determinate the state of circuit because it is in state 0
#if in one qubitcircuit I want to change the 0 state to 1
putstate(qc,1)

}
