% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paulimatrices.R
\name{empty_dim}
\alias{empty_dim}
\title{Return a list of 2 vectors with <a> empty strings in the first and <b> empty strings in the secons}
\usage{
empty_dim(a, b)
}
\arguments{
\item{a}{- integer}

\item{b}{- integer}
}
\value{
list
}
\description{
Return a list of 2 vectors with <a> empty strings in the first and <b> empty strings in the secons
}
\examples{
I <- matrix(c(1,0,0,1), ncol=2, nrow=2, dimnames=empty_dim(2,2))
I

}
\seealso{
\link{pauli.getrx}, \link{pauli.getry}, \link{pauli.getrz},
}
