% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encryption.R
\name{encrypt_message}
\alias{encrypt_message}
\title{Encrypt a message}
\usage{
encrypt_message(message, S)
}
\arguments{
\item{message}{- text}

\item{S}{- a singular 4x4 matrix}
}
\value{
Encrypted message
}
\description{
Encrypt a message
}
\examples{
S <- array(as.integer(c(2,1,3,1,3,1,3,2,4,2,3,3,1,1,1,1)), dim=c(4,4), dimnames=empty_dim(4,4))
Encryption <- encrypt_message("Hello", S)
}
\seealso{
\link{decrypt_message}
}
