% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaryconverters.R
\name{gray_to_decimal}
\alias{gray_to_decimal}
\title{Gray code to decimal conversion}
\usage{
gray_to_decimal(gray)
}
\arguments{
\item{gray}{- Number in gray format}
}
\value{
Given number converted from gray to decimal
}
\description{
Gray code to decimal conversion
}
\examples{
gray_to_decimal(c(1,0))  # returns 3
}
\seealso{
\link{decimal_to_gray}, \link{gray_to_binary}
}
