% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{integer_part}
\alias{integer_part}
\title{Get integer part of a number}
\usage{
integer_part(X)
}
\arguments{
\item{X}{- number}
}
\value{
Integer part of X
}
\description{
Get integer part of a number
}
\examples{
integer_part(4.5)  # E(4.5)  = 4
integer_part(-4.5) # E(-4.5) = -5

}
\seealso{
\link{matrices.mod}, \link{matrices.integerpart}
}
