% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{matrices.determinant}
\alias{matrices.determinant}
\title{Determinant of a real or complex matrix}
\usage{
matrices.determinant(A)
}
\arguments{
\item{A}{: a square matrix}
}
\value{
Determinant of the given matrix
}
\description{
Use this function to compute the determinant of square matrix
}
\examples{
S <- array(c(1:10, 1, 5, 5, 3, 4, 1), dim=c(4,4))  # Exemple with a 4x4 matrix
matrices.determinant(S)

}
\seealso{
\link{matrices.reduce}, \link{matrices.inverse}
}
