% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{matrices.identity}
\alias{matrices.identity}
\title{Build Identity matrix}
\usage{
matrices.identity(M, N = M)
}
\arguments{
\item{M}{: First dimension}

\item{N}{: Second dimension}
}
\value{
Identity matrix of space MxN
}
\description{
Build Identity matrix
}
\examples{
matrices.identity(3)
matrices.identity(4, 3)
}
