% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{matrices.integerpart}
\alias{matrices.integerpart}
\title{matrices.E or matrices.integer_part computes integer part of all elements of the matrix}
\usage{
matrices.integerpart(matrix)
}
\arguments{
\item{matrix}{- Given numeric matrix to be computed}
}
\value{
Matrix containing integer part of elements of given matrix in exact position
}
\description{
matrices.E or matrices.integer_part computes integer part of all elements of the matrix
}
\examples{
A <- array(rnorm(16, 8, 56), dim=c(4,4))
matrices.integerpart(A)

}
\seealso{
\link{matrices.mod}
}
