% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{matrices.inverse}
\alias{matrices.inverse}
\title{Computes the inverse of an invertible square matrix}
\usage{
matrices.inverse(A)
}
\arguments{
\item{A}{- square matrix}
}
\value{
Inverse of the matrix A
}
\description{
Computes the inverse of an invertible square matrix
}
\examples{
A <- array(c(4,8,9,2,7,9,2,6,0), dim=c(3,3))
matrices.inverse(A)

}
\seealso{
\link{matrices.reduce}
}
