% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{matrices.linearsolve}
\alias{matrices.linearsolve}
\title{Solve systems of linear equations}
\usage{
matrices.linearsolve(A, B)
}
\arguments{
\item{A}{A square matrix with coefficients of variables}

\item{B}{The matrix of results}
}
\value{
The matrix of solution for each result
}
\description{
Solve systems of linear equations
}
\examples{


# Example 1 - with a vector of results
A <- array(c(1,3,2,1), dim=c(2,2))
B <- c(1, 2)
matrices.linearsolve(A, B)

# Example 2 - with a matrix of results
B <- array(c(1, 2, 2, 1, 5, 0), dim=c(2,3))
matrices.linearsolve(A, B)
}
