% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{matrices.mod}
\alias{matrices.mod}
\title{computes de modulus in a given base of each element of a matrix}
\usage{
matrices.mod(A, number)
}
\arguments{
\item{A}{- a numeric vector or matrix}

\item{number}{- the given base used for operation}
}
\value{
Matrix or vector with all elements in mod[number]
}
\description{
computes de modulus in a given base of each element of a matrix
}
\examples{
A <- 1:6
matrices.mod(A, 3)
B <- array(c(2,3,4,6,7,-1,-1,4), dim=c(4,2))
matrices.mod(B, 27)

}
\seealso{
\link{matrices.E}
}
