% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{matrices.product}
\alias{matrices.product}
\title{Computes matrix product of two matrices}
\usage{
matrices.product(A, B)
}
\arguments{
\item{A}{- matrix (M x N)}

\item{B}{- matrix (N x P)}
}
\value{
matrix (M x P) which is the product of and and B
}
\description{
Computes matrix product of two matrices
}
\examples{
A <- matrix(1:4, nrow=1, ncol=4)
B <- matrix(1:16, nrow=4, ncol=4)
matrices.product(A, B)

}
\seealso{
\link{matrices.power}
}
