% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{matrices.reduce}
\alias{matrices.reduce}
\title{Gauss-Jordan reduction of matrix}
\usage{
matrices.reduce(A)
}
\arguments{
\item{A}{- matrix}
}
\value{
Reduced matrix
}
\description{
Gauss-Jordan reduction of matrix
}
\details{
If the output is an Identity matrix then the determinant of this matrix is non null, hence inversible
}
\examples{
A <- array(c(4,8,9,2,7,9,2,6,0), dim=c(3,3))
matrices.reduce(A)
}
