% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{matrices.trace}
\alias{matrices.trace}
\title{Trace of a matrix}
\usage{
matrices.trace(A)
}
\arguments{
\item{A}{: A square matrix}
}
\value{
Integer
}
\description{
Product of elements of the diagonal
}
\examples{
A <- matrix(c(1:4), ncol=2, nrow=2)
matrices.trace(A)
}
