% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurement.R
\name{measurement}
\alias{measurement}
\title{measurement}
\usage{
measurement(a, b)
}
\description{
measurement
}
\examples{
# to apply hadamard to the first qubit of a circuit called qc

qc$q_1<-Hadamard(qc$q_1)
}
