% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paulimatrices.R
\name{pauli.getrx}
\alias{pauli.getrx}
\title{Getter method for Pauli rx spin matrix}
\usage{
pauli.getrx()
}
\value{
Pauli rx spin matrix
}
\description{
Getter method for Pauli rx spin matrix
}
\examples{
pauli.getrx()
}
