% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paulimatrices.R
\name{pauli.getry}
\alias{pauli.getry}
\title{Getter method for Pauli rx spin matrix}
\usage{
pauli.getry()
}
\value{
Pauli ry spin matrix
}
\description{
Getter method for Pauli rx spin matrix
}
\examples{
pauli.getry()
}
