% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paulimatrices.R
\name{pauli.getrz}
\alias{pauli.getrz}
\title{Getter method for Pauli rx spin matrix}
\usage{
pauli.getrz()
}
\value{
Pauli r\ spin matrix
}
\description{
Getter method for Pauli rx spin matrix
}
\examples{
pauli.getrz()
}
