% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmeasurement.R
\name{plotMeasurement}
\alias{plotMeasurement}
\title{Plot measurement}
\usage{
plotMeasurement(a)
}
\description{
Plot measurement
}
\examples{
# to apply hadamard to the first qubit of a circuit called qc

plotMeasurement(qc)

}
