% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantumcircuit.R
\name{quantumcircuit}
\alias{quantumcircuit}
\title{quantumcircuit}
\usage{
quantumcircuit(nqubit)
}
\description{
quantumcircuit
}
\examples{
# to create a quantum circuit of two qubits called qc

qc<-quantumcircuit(2)

}
