% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codetable.R
\name{ravel_matrix}
\alias{ravel_matrix}
\title{Redimesion a MxN matrix into a 1x(M*N) matrix by columns}
\usage{
ravel_matrix(matrix)
}
\arguments{
\item{matrix}{Matrix to be ravelled}
}
\value{
Ravelled matrix
}
\description{
Redimesion a MxN matrix into a 1x(M*N) matrix by columns
}
\examples{
A <- array(c(1,2,3,4), dim=c(2,2))
ravel_matrix(A)  # return c(1,3,2,4)

}
\seealso{
\link{codetable1.numerise_char}, \link{codetable1.numerise_char.as_matrix}
}
