% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encryption.R
\name{send_encrypted_message}
\alias{send_encrypted_message}
\title{Send an encrypted message to a peer receiver on a network}
\usage{
send_encrypted_message(message, S, address = "localhost", PORT = "27015")
}
\arguments{
\item{message}{: String message to be sent}

\item{S}{: Singular matrix for encryption}

\item{address}{: Address of peer receiver}

\item{PORT}{: PORT at which receiver listens}
}
\value{
1 if succeeded else 0
}
\description{
The receiver hosts a server at an address and you want to send an encrypted message to the server. The message is crypted and sent. A prerequisite is that the recepter must be waiting the message
}
\examples{
S <-  array(as.integer(c(2,1,3,1,3,1,3,2,4,2,3,3,1,1,1,1)), dim=c(4,4), dimnames=empty_dim(4,4))
send_encrypted_message("Catch It", S, "localhost", "27014")   # This will work only if peer is waiting message at that location
}
\seealso{
\link{receive_decrypt_message}
}
