% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{send_message}
\alias{send_message}
\title{Send a message to a network address}
\usage{
send_message(message, address, PORT)
}
\arguments{
\item{message}{: String}

\item{address}{: IP address}

\item{PORT}{: Port to communicate}
}
\value{
Status of communication
}
\description{
This function enables to send a message to a particular address, on a particular port. This requires that a server is listening to the port
}
\examples{
message <- "The Open QUANTUM Security"
address <- "localhost"
PORT    <- "27015"
send_message(message, address, PORT)

}
