% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codetable.R
\name{sentence}
\alias{sentence}
\title{Merge all the elements of a vector into a string}
\usage{
sentence(vector, sep = "")
}
\arguments{
\item{vector}{- vector containing elements to be concatenated into a string}

\item{sep}{- seperator}
}
\value{
String of all the vector elements merged
}
\description{
Merge all the elements of a vector into a string
}
\examples{
sentence(c("H", "e", "l", "l", "o"))
}
