% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codetable.R
\name{split}
\alias{split}
\title{Split a String into a character vector}
\usage{
split(string, sep = "~")
}
\arguments{
\item{string}{: text to be parsed}

\item{sep}{: seperator}
}
\value{
Character vector containg the elements splitted according to the seperator
}
\description{
Split a String into a character vector
}
\examples{
split("Hello world ! Guess my name", sep=" ")
}
