% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{wait_message}
\alias{wait_message}
\title{Wait message from a port}
\usage{
wait_message(PORT)
}
\arguments{
\item{PORT}{: PORT to listen}
}
\value{
received message
}
\description{
Listen to a port on your machine to wait for a message. Ensure the port is free. The sender will communicate to this port
}
\examples{
wait_message(27015)

}
\seealso{
\link{send_message}
}
